% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmanager-package.R
\docType{package}
\name{dbmanager-package}
\alias{dbmanager-package}
\alias{_PACKAGE}
\title{dbmanager: A connection manager for databases and compute engines.}
\description{
Provides a consistent interface for managing 
       connections to database and compute engines in R.
}
\author{
\strong{Maintainer}: Ernest Omane-Kodie \email{eokodie@gmail.com}

}
\keyword{internal}
