% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monetdb_lite.R
\docType{data}
\name{MonetDBLite}
\alias{MonetDBLite}
\title{Manage MonetDBLite connections}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
MonetDBLite
}
\arguments{
\item{db_dir}{DB directory.}
}
\description{
This is a class for managing MonetDBLite connections.

open a MonetDBLite connection
}
\examples{
\dontrun{
library(dbmanager)
monetdb <- MonetDBLite$new(tempdir())
monetdb$connected_database
monetdb$tables
monetdb$close()
} 

}
\keyword{datasets}
