% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\docType{data}
\name{MySQL}
\alias{MySQL}
\title{mysql}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
MySQL
}
\arguments{
\item{db_name}{Database name.}

\item{host}{Host name.}

\item{user}{User name.}

\item{password}{Password.}
}
\value{
A new `Pool` object.
}
\description{
This class is a class for managing MySQL connections.

open a MySQL connection
}
\examples{
\dontrun{
library(dbmanager)
mysql <- MySQL$new(db_name, host, user, password)
mysql$available_databases
mysql$connected_database
mysql$tables
mysql$close()
} 

}
\keyword{datasets}
