% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\docType{data}
\name{PostgreSQL}
\alias{PostgreSQL}
\title{MySQL}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
PostgreSQL
}
\arguments{
\item{db_name}{Database name.}

\item{host}{Host name.}

\item{user}{User name.}

\item{password}{Password.}
}
\value{
A new `Pool` object.
}
\description{
This is a class for managing PostgreSQL connections.

open a PostgreSQL connection
}
\examples{
\dontrun{
library(dbmanager)
postgres <- PostgreSQL$new(db_name, host, user, password)
postgres$available_databases
postgres$connected_database
postgres$tables
postgres$close()
} 

}
\keyword{datasets}
