% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\docType{data}
\name{SQLite}
\alias{SQLite}
\title{Manage SQLite connections}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SQLite
}
\arguments{
\item{host}{DB host}
}
\description{
This is a class for managing SQLite connections.

open a MonetDBLite connection
}
\examples{
\dontrun{
library(dbmanager)
db <- SQLite$new(host = ":memory:")
db$connected_database
db$tables
db$close()
} 

}
\keyword{datasets}
