% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_server.R
\docType{data}
\name{SQLServer}
\alias{SQLServer}
\title{Manage SQL Server connections}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SQLServer
}
\arguments{
\item{db_name}{Database name.}

\item{server_url}{Server URL.}

\item{user}{User name.}

\item{password}{Password.}

\item{port}{Port number}
}
\value{
A new `Pool` object.
}
\description{
This is a class for managing SQL Server connections.

open a SQLServer connection
}
\examples{
\dontrun{
library(dbmanager)
sql_server <- SQLServer$new(db_name, server_url, username, password, port)
sql_server$available_databases
sql_server$connected_database
sql_server$tables
sql_server$close()
} 

}
\keyword{datasets}
