% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\docType{data}
\name{DatabaseManager}
\alias{DatabaseManager}
\title{Manage Databases}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
DatabaseManager
}
\arguments{
\item{db_name}{Database name.}

\item{host}{Host name.}

\item{user}{User name.}

\item{password}{Password.}

\item{name}{A character string specifying the unquoted DBMS table name.}

\item{value}{a data.frame (or coercible to data.frame).}
}
\value{
A DatabaseManager template.
Create a new `DatabaseManager` object.

A new `DatabaseManager` object.
open a database connection
}
\description{
This is a superclass for managing database connections.

Create a table
}
\section{Methods}{

\describe{
  \item{\code{open}}{open a database connection}
  \item{\code{close}}{close database connection}
  \item{\code{show_databases}}{show available databases}
  \item{\code{create_new_db}}{create a new database}
  \item{\code{create_table}}{create a table}
}
}

\keyword{datasets}
