% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{dbmanager-package}
\alias{dbmanager}
\alias{dbmanager-package}
\title{dbmanager: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ernest Omane-Kodie \email{eokodie@gmail.com}

}
